drivers:dac:dac_demo: Use submit function instead of write_dev
Use submit function instead of write_dev for buffer writings.
Move dac_submit_samples and dac_demo_trigger_handler
implementations to iio_dac_demo.c.
